move types to api clients folder
